requirements.txt
# Core dependencies
python-dotenv==1.2.1
python-telegram-bot==20.6
httpx==0.25.2
autogen-ext==0.7.5
autogen-agentchat==0.7.5
pyautogen==0.2.35
requests==2.32.5
